WIP moved filterMutants logs to status reporter
