+++ src/Stryker.Core/Stryker.Core/Reporters/StatusReporter.cs
@@ -1,6 +1,8 @@
 using System;
+using Crayon;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Stryker.Core.Mutants;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
@@ -24,7 +26,17 @@ namespace Stryker.Core.Reporters
 
         public void OnMutantsCreated(IReadOnlyInputComponent reportComponent)
         {
-            // This reporter does not report during the testrun
+            var notRunMutantsWithResultStatusReason = reportComponent.ReadOnlyMutants
+                .Where(m => m.ResultStatus == MutantStatus.NotRun && !string.IsNullOrEmpty(m.ResultStatusReason))
+                .GroupBy(x => x.ResultStatusReason);
+
+            foreach (var notRunMutantReason in notRunMutantsWithResultStatusReason)
+            {
+                _consoleWriter.WriteLine(LeftPadAndFormatForMutantCount(notRunMutantReason.Count(), "mutants will be tested because: {1}"));
+            }
+
+            var notRunCount = reportComponent.ReadOnlyMutants.Count(m => m.ResultStatus == MutantStatus.NotRun);
+            _consoleWriter.WriteLine(LeftPadAndFormatForMutantCount(notRunCount, "total mutants will be tested"));
         }
 
         public void OnMutantTested(IReadOnlyMutant result)
@@ -34,7 +46,36 @@ namespace Stryker.Core.Reporters
 
         public void OnStartMutantTestRun(IEnumerable<IReadOnlyMutant> mutantsToBeTested, IEnumerable<TestDescription> testDescriptions)
         {
-            throw new System.NotImplementedException();
+            var skippedMutantGroups = mutantsToBeTested.GroupBy(x => new { x.ResultStatus, x.ResultStatusReason }).OrderBy(x => x.Key.ResultStatusReason);
+
+            foreach (var skippedMutantGroup in skippedMutantGroups)
+            {
+                _consoleWriter.WriteLine(FormatStatusReasonLogString(skippedMutantGroup.Count(), skippedMutantGroup.Key.ResultStatus));
+            }
+
+            if (mutantsToBeTested.Any())
+            {
+                _consoleWriter.WriteLine(LeftPadAndFormatForMutantCount(mutantsToBeTested.Count(), "total mutants are skipped for the above mentioned reasons"));
+            }
+
+        }
+
+        private string FormatStatusReasonLogString(int mutantCount, MutantStatus resultStatus)
+        {
+            // Pad for status CompileError length
+            var padForResultStatusLength = 13 - resultStatus.ToString().Length;
+
+            var formattedString = LeftPadAndFormatForMutantCount(mutantCount, "mutants got status {1}.");
+            formattedString += "Reason: {2}".PadLeft(11 + padForResultStatusLength);
+
+            return formattedString;
+        }
+
+        private string LeftPadAndFormatForMutantCount(int mutantCount, string logString)
+        {
+            // Pad for max 5 digits mutant amount
+            var padLengthForMutantCount = 5 - mutantCount.ToString().Length;
+            return "{0} " + logString.PadLeft(logString.Length + padLengthForMutantCount);
         }
     }
 }
